+++ core/components/cronmanager/docs/changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - TBA
+### Changed
+- Fix a default sort issue
+### Added
+- Output the run date in webcronjob mode
+- Log request parameter issues in webcronjob mode
+
 ## [1.2.0] - 2019-09-10
 ### Changed
 - Removed modAction Support